Added the images and made the header
Filled the header and downloaded the images.
Now it's all a matter of correctly positioning the elements.
In the nav the three rows have the same height
which is unwanted.
The icons in the articles cannot be aligned to the
right using display: flex like that.
Look for how to get semicirlces to the side of the buttons
in the header. Right now the curvature starts from the center
of the button.