FIXED THE REPOSITORY AND STYLED THE NAV AND HEADER
There has been the usual problem with HEAD, maybe its
because we are saving on the hard disk.
Styled nav and header. Need to figure out how to set
the row height so that it is the same for the first
and next rows (auto rows). Next is the side page.